changes stack setup and adds tabs
