+++ app/_layout.tsx
@@ -3,25 +3,23 @@ import { Stack } from "expo-router";
 import React, { useState } from "react";
 
 export default function RootLayout() {
-	interface User {
-		username?: string;
-		password?: string;
-		email?: string;
-		user_id?: string;
-		decks?: number[] | null;
-		_id?: string;
-	}
+  interface User {
+    username?: string;
+    password?: string;
+    email?: string;
+    user_id?: string;
+    decks?: number[] | null;
+    _id?: string;
+  }
 
-	const [userDetails, setUserDetails] = useState<User>({ username: "ana" });
+  const [userDetails, setUserDetails] = useState<User>({ username: "ana" });
 
-	return (
-		<UserContext.Provider value={{ userDetails, setUserDetails }}>
-			<Stack>
-				<Stack.Screen name="index" />
-				<Stack.Screen name="Login" />
-				<Stack.Screen name="Signup" />
-				<Stack.Screen name="home" />
-			</Stack>
-		</UserContext.Provider>
-	);
+  return (
+    <UserContext.Provider value={{ userDetails, setUserDetails }}>
+      <Stack>
+        <Stack.Screen name="(tabs)" />
+        {/* remove (tabs) display - see tutorial */}
+      </Stack>
+    </UserContext.Provider>
+  );
 }
