+++ components-example/ThemedText.tsx
